requirements.txt
pypdf2
spacy_langdetect
deep_translator
sentence-transformers
faiss-cpu
transformers
tensorflow
tf-keras
spacy
bertopic
gdown
nltk
numpy==2.0.2
